Removing batch_size attribute and made the Crawler extensible
