+++ data-prepper-plugins/saas-source-plugins/atlassian-commons/src/main/java/org/opensearch/dataprepper/plugins/source/atlassian/AtlassianSourceConfig.java
@@ -21,8 +21,6 @@ import java.util.List;
 @Getter
 public abstract class AtlassianSourceConfig implements CrawlerSourceConfig {
 
-    private static final int DEFAULT_BATCH_SIZE = 50;
-
     /**
      * Jira account url
      */
@@ -35,12 +33,6 @@ public abstract class AtlassianSourceConfig implements CrawlerSourceConfig {
     @JsonProperty("authentication")
     @Valid
     protected AuthenticationConfig authenticationConfig;
-
-    /**
-     * Batch size for fetching tickets
-     */
-    @JsonProperty("batch_size")
-    protected int batchSize = DEFAULT_BATCH_SIZE;
 
 
     /**
