+++ data-prepper-plugins/saas-source-plugins/confluence-source/src/main/java/org/opensearch/dataprepper/plugins/source/confluence/ConfluenceSourceConfig.java
@@ -19,8 +19,6 @@ import org.opensearch.dataprepper.plugins.source.source_crawler.base.CrawlerSour
 @Getter
 public class ConfluenceSourceConfig extends AtlassianSourceConfig implements CrawlerSourceConfig {
 
-    private static final int DEFAULT_BATCH_SIZE = 50;
-
     /**
      * Filter Config to filter what tickets get ingested
      */
