+++ data-prepper-plugins/saas-source-plugins/confluence-source/src/test/java/org/opensearch/dataprepper/plugins/source/confluence/configuration/PageTypeConfigTest.java
@@ -35,7 +35,7 @@ class PageTypeConfigTest {
         pageTypeConfig.exclude.add("page");
         pageTypeConfig.exclude.add("blogpost");
 
-        assertFalse(pageTypeConfig.isValidPageType());
+        assertFalse(pageTypeConfig.noOverlapBetweenIncludeAndExclude());
     }
 
     @Test
