+++ data-prepper-plugins/saas-source-plugins/source-crawler/src/main/java/org/opensearch/dataprepper/plugins/source/source_crawler/base/CrawlerSourceConfig.java
@@ -7,8 +7,6 @@ public interface CrawlerSourceConfig {
 
     int DEFAULT_NUMBER_OF_WORKERS = 1;
 
-    int getBatchSize();
-
     /**
      * Boolean to indicate if acknowledgments enabled for this source
      *
